Add unhandledrejection option
I tried using a middleware to log all unhandled promises, and although it's clear now that I've stopped and thought about it, it took me a little while to realize that the middleware fired before the action creators did (so it couldn't know what was handled and what wasn't). My solution was to use a browser / Promises library "unhandledrejection" event.

This PR adds a bit of info about these alternatives and also tries to clarify a bit regarding redux-promise-middleware ("the middleware") versus custom error-handling middleware that the developer adds ("a middleware").

Feedback welcome.